value type is corrected

+++ src/BmiService.java
@@ -1,6 +1,6 @@
 public class BmiService {
-    public double calculate(int bodyMass, double height) {
-        double result = bodyMass / Math.pow(height, 2);
+    public int calculate(int bodyMass, double height) {
+        int result = (int) (bodyMass / Math.pow(height, 2));
         return result;
     }
 }
